requirements.txt
asgiref==3.5.2
Django==4.1.1
django-debug-toolbar==3.7.0
django-decouple==2.1
sqlparse==0.4.3
tzdata==2022.4
